+++ Changes.md
@@ -12,6 +12,7 @@ Improvements
 - Annotations :
   - Added support for `{plug}` value substitutions in node annotations.
   - Added <kbd>Ctrl</kbd> + <kbd>Enter</kbd> keyboard shortcut to annotation dialogue. This applies the annotation and closes the dialogue.
+- LightEditor : Added support for Arnold light blockers.
 
 Fixes
 -----